+++ src/redux/actions.ts
@@ -1,7 +1,7 @@
 export const SET_ACCESS_TOKEN = 'SET_ACCESS_TOKEN';
 export const SET_REFRESH_TOKEN = 'SET_REFRESH_TOKEN';
 
-export const SET_USER_NAME = 'SET_USER_NAME';
+export const SET_USERNAME = 'SET_USERNAME';
 
 export const setAccessToken = (accessToken: string | null) => (dispatch: any) => {
   dispatch({ type: SET_ACCESS_TOKEN, payload: accessToken });
@@ -11,6 +11,6 @@ export const setRefreshToken = (refreshToken: string | null) => (dispatch: any) 
   dispatch({ type: SET_REFRESH_TOKEN, payload: refreshToken });
 };
 
-export const setUserName = (username: string | null) => (dispatch: any) => {
-  dispatch({ type: SET_USER_NAME, payload: username });
+export const setUsername = (username: string | null) => (dispatch: any) => {
+  dispatch({ type: SET_USERNAME, payload: username });
 };