+++ src/redux/reducers.ts
@@ -1,4 +1,4 @@
-import { SET_ACCESS_TOKEN, SET_REFRESH_TOKEN, SET_USER_NAME } from './actions';
+import { SET_ACCESS_TOKEN, SET_REFRESH_TOKEN, SET_USERNAME } from './actions';
 
 const initialState = {
   accessToken: null,
@@ -12,8 +12,8 @@ export default function userReducer(state = initialState, action: any) {
       return { ...state, accessToken: action.payload };
     case SET_REFRESH_TOKEN:
       return { ...state, refreshToken: action.payload };
-    case SET_USER_NAME:
-      return { ...state, userName: action.payload };
+    case SET_USERNAME:
+      return { ...state, username: action.payload };
     default:
       return state;
   }