+++ src/screens/CourtScreen.tsx
@@ -1,15 +1,15 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { Button, Heading, useColorMode, View } from 'native-base';
 import { useDispatch } from 'react-redux';
-import { RootTabScreenProps } from '../types';
 import * as SecureStore from 'expo-secure-store';
 import { setAccessToken, setRefreshToken } from '../redux/actions';
+import { HomeStackScreenProps } from '../navigation/types';
 
-export default function CourtScreen({ navigation }: RootTabScreenProps<'Court'>) {
+export default function CourtScreen({ navigation }: HomeStackScreenProps<'Court'>) {
   const { colorMode, toggleColorMode } = useColorMode();
 
   const dispatch = useDispatch();
-
+  // BUG after navigating back from chat on checkNetwork() and color mode change the app renavigate to court
   const checkNetwork = () => {
     console.log('get protected route');
     axios
@@ -18,19 +18,12 @@ export default function CourtScreen({ navigation }: RootTabScreenProps<'Court'>)
         console.log(response.data);
       })
       .catch((error: AxiosError) => {
-        console.log("can't react protect route", error.response?.status);
+        console.log("can't react protect route", error.message);
       });
   };
 
-  const failedReq = () => {
-    axios
-      .get('/404error')
-      .then((response: AxiosResponse) => {
-        console.log(response.data);
-      })
-      .catch((error: AxiosError) => {
-        console.log('404 error');
-      });
+  const navigateToChat = () => {
+    navigation.navigate('Chat');
   };
 
   const logout = async () => {
@@ -50,8 +43,8 @@ export default function CourtScreen({ navigation }: RootTabScreenProps<'Court'>)
       <Button mt={5} onPress={checkNetwork}>
         Check token
       </Button>
-      <Button mt={5} onPress={failedReq}>
-        404 request
+      <Button mt={5} onPress={navigateToChat} bg={'green.600'}>
+        Chat
       </Button>
       <Button mt={5} onPress={logout}>
         Logout